+++ ionoscloud/data_source_servers_test.go
@@ -1,4 +1,4 @@
-//go:build compute || all || servers
+//go:build compute || all || server
 
 package ionoscloud
 
@@ -17,7 +17,6 @@ import (
 const serverTestResource2 = constant.ServerTestResource + "2"
 
 func TestAccDataSourceServersBasic(t *testing.T) {
-	t.Skip("problem with 500 error thrown by backend")
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)